--- apps/web/src/app/partner/mypage/_shared/model/types.ts
@@ -1,10 +1,0 @@
-import type { z } from "zod";
-import type {
-  notificationSettingsSchema,
-  updateNotificationSettingsBodySchema,
-} from "./notification-settings.schema";
-
-export type NotificationSettings = z.infer<typeof notificationSettingsSchema>;
-export type UpdateNotificationSettingsBody = z.infer<
-  typeof updateNotificationSettingsBodySchema
->;
