+++ apps/web/src/shared/api/get-notification-settings.ts
@@ -1,7 +1,7 @@
 import { API_BASE_URL } from "@/shared/api/config";
 import { fetchJson } from "@/shared/api/fetch-json";
-import { notificationSettingsSchema } from "../model/notification-settings.schema";
-import type { NotificationSettings } from "../model/types";
+import { notificationSettingsSchema } from "@/shared/model/notification-settings.schema";
+import type { NotificationSettings } from "@/shared/model/notification-settings.schema";
 
 export function getNotificationSettings(): Promise<NotificationSettings> {
   return fetchJson(
