+++ apps/web/src/shared/api/use-notification-settings.ts
